+++ example/lib/main.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_map_marker_popup_example/popup_manager_example.dart';
 import 'package:flutter_map_marker_popup_example/popup_option_controls.dart';
 import 'package:flutter_map_marker_popup_example/simple_map_with_popups.dart';
+
+import 'popup_outside_of_map.dart';
 
 void main() => runApp(const MyApp());
 
@@ -16,8 +17,8 @@ class MyApp extends StatelessWidget {
       home: const PopupOptionControls(),
       routes: <String, WidgetBuilder>{
         PopupOptionControls.route: (context) => const PopupOptionControls(),
-        SimpleMapWithPopups.route: (context) => SimpleMapWithPopups(),
-        PopupManagerExample.route: (context) => PopupManagerExample(),
+        SimpleMapWithPopups.route: (context) => const SimpleMapWithPopups(),
+        PopupOutsideOfMap.route: (context) => const PopupOutsideOfMap(),
       },
     );
   }