+++ lib/extension_api.dart
@@ -7,3 +7,4 @@ export 'src/popup_container/simple_popup_container.dart';
 export 'src/popup_controller.dart';
 export 'src/popup_layer.dart';
 export 'src/popup_snap.dart';
+export 'src/popup_state_wrapper.dart';