+++ lib/flutter_map_marker_popup.dart
@@ -6,4 +6,6 @@ export 'src/popup_animation.dart';
 export 'src/popup_controller.dart';
 export 'src/popup_marker_layer_options.dart';
 export 'src/popup_marker_layer_widget.dart';
+export 'src/popup_scope.dart';
 export 'src/popup_snap.dart';
+export 'src/popup_state.dart';